refactor: move flags related to revoke to flags.go
