+++ cmd/cmd_revoke.go
@@ -4,16 +4,9 @@ import (
 	"context"
 	"log/slog"
 
-	"github.com/go-acme/lego/v5/acme"
 	"github.com/go-acme/lego/v5/cmd/internal/storage"
 	"github.com/go-acme/lego/v5/log"
 	"github.com/urfave/cli/v3"
-)
-
-// Flag names.
-const (
-	flgKeep   = "keep"
-	flgReason = "reason"
 )
 
 func createRevoke() *cli.Command {
@@ -23,30 +16,6 @@ func createRevoke() *cli.Command {
 		Action: revoke,
 		Flags:  createRevokeFlags(),
 	}
-}
-
-func createRevokeFlags() []cli.Flag {
-	flags := CreateBaseFlags()
-
-	flags = append(flags,
-		&cli.BoolFlag{
-			Name:    flgKeep,
-			Aliases: []string{"k"},
-			Usage:   "Keep the certificates after the revocation instead of archiving them.",
-		},
-		&cli.UintFlag{
-			Name: flgReason,
-			Usage: "Identifies the reason for the certificate revocation." +
-				" See https://www.rfc-editor.org/rfc/rfc5280.html#section-5.3.1." +
-				" Valid values are:" +
-				" 0 (unspecified), 1 (keyCompromise), 2 (cACompromise), 3 (affiliationChanged)," +
-				" 4 (superseded), 5 (cessationOfOperation), 6 (certificateHold), 8 (removeFromCRL)," +
-				" 9 (privilegeWithdrawn), or 10 (aACompromise).",
-			Value: acme.CRLReasonUnspecified,
-		},
-	)
-
-	return flags
 }
 
 func revoke(ctx context.Context, cmd *cli.Command) error {
